Moved particle spectra over
