+++ src/synthesizer/particle/galaxy.py
@@ -192,67 +192,6 @@ class Galaxy(BaseGalaxy):
         self.gas = Gas(masses, metals, **kwargs)
         self.calculate_integrated_gas_properties()
 
-    def _prepare_sed_args(self, grid, fesc, spectra_type, mask=None):
-        """
-        A method to prepare the arguments for SED computation with the C
-        functions.
-
-        Args:
-            grid (Grid)
-                The SPS grid object to extract spectra from.
-            fesc (float)
-                The escape fraction.
-            spectra_type (str)
-                The type of spectra to extract from the Grid. This must match a
-                type of spectra stored in the Grid.
-            mask (bool)
-                A mask to be applied to the stars. Spectra will only be computed
-                and returned for stars with True in the mask.
-        """
-
-        if mask is None:
-            mask = np.ones(self.stars.nparticles, dtype=bool)
-
-        # Set up the inputs to the C function.
-        grid_props = [
-            np.ascontiguousarray(grid.log10age, dtype=np.float64),
-            np.ascontiguousarray(np.log10(grid.metallicity), dtype=np.float64),
-        ]
-        part_props = [
-            np.ascontiguousarray(self.stars.log10ages[mask], dtype=np.float64),
-            np.ascontiguousarray(self.stars.log10metallicities[mask], dtype=np.float64),
-        ]
-        part_mass = np.ascontiguousarray(
-            self.stars._initial_masses[mask], dtype=np.float64
-        )
-        npart = np.int32(part_mass.size)
-        nlam = np.int32(grid.spectra[spectra_type].shape[-1])
-
-        # Slice the spectral grids and pad them with copies of the edges.
-        grid_spectra = np.ascontiguousarray(grid.spectra[spectra_type], np.float64)
-
-        # Get the grid dimensions after slicing what we need
-        grid_dims = np.zeros(len(grid_props) + 1, dtype=np.int32)
-        for ind, g in enumerate(grid_props):
-            grid_dims[ind] = len(g)
-        grid_dims[ind + 1] = nlam
-
-        # Convert inputs to tuples
-        grid_props = tuple(grid_props)
-        part_props = tuple(part_props)
-
-        return (
-            grid_spectra,
-            grid_props,
-            part_props,
-            part_mass,
-            fesc,
-            grid_dims,
-            len(grid_props),
-            npart,
-            nlam,
-        )
-
     def calculate_black_hole_metallicity(self, default_metallicity=0.012):
         """
         Calculates the metallicity of the region surrounding a black hole. This
@@ -410,153 +349,6 @@ class Galaxy(BaseGalaxy):
             force_loop,
         )
 
-    def generate_lnu(
-        self, grid, spectra_name, fesc=0.0, young=False, old=False, verbose=False
-    ):
-        """
-        Generate the luminosity for a given grid key spectra for all
-        stars in this galaxy object. Can optionally apply masks.
-
-        Base class for :func:`~particle.ParticleGalaxystars.get_spectra_incident`
-        and other related methods
-
-        Args:
-            grid (obj):
-                spectral grid object
-            spectra_name (string):
-                name of the target spectra inside the grid file
-            fesc (float):
-                fraction of stellar emission that escapes unattenuated from
-                the birth cloud (defaults to 0.0)
-            young (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for young star particles
-            old (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for old star particles
-            verbose (bool):
-                Flag for verbose output
-
-        Returns:
-            numpy array of integrated spectra in units of (erg / s / Hz)
-        """
-
-        # Ensure we have a total key in the grid. If not error.
-        if spectra_name not in list(grid.spectra.keys()):
-            raise exceptions.MissingSpectraType(
-                "The Grid does not contain the key '%s'" % spectra_name
-            )
-
-        # get particle age masks
-        mask = self._get_masks(young, old)
-
-        if np.sum(mask) == 0:
-            if verbose:
-                print("Age mask has filtered out all particles")
-
-            return np.zeros(len(grid.lam))
-
-        from ..extensions.integrated_spectra import compute_integrated_sed
-
-        # Prepare the arguments for the C function.
-        args = self._prepare_sed_args(
-            grid, fesc=fesc, spectra_type=spectra_name, mask=mask
-        )
-
-        # Get the integrated spectra in grid units (erg / s / Hz)
-        spec = compute_integrated_sed(*args)
-
-        return spec
-
-    def generate_particle_lnu(
-        self, grid, spectra_name, fesc=0.0, young=False, old=False, verbose=False
-    ):
-        """
-        Generate the luminosity for a given grid key spectra for all
-        stars in this galaxy object. Can optionally apply masks.
-
-        Base class for :func:`~particle.ParticleGalaxystars.get_spectra_incident`
-        and other related methods
-
-        Args:
-            grid (obj):
-                spectral grid object
-            spectra_name (string):
-                name of the target spectra inside the grid file
-            fesc (float):
-                fraction of stellar emission that escapes unattenuated from
-                the birth cloud (defaults to 0.0)
-            young (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for young star particles
-            old (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for old star particles
-            verbose (bool):
-                Flag for verbose output
-
-        Returns:
-            numpy array of integrated spectra in units of (erg / s / Hz)
-        """
-
-        # Ensure we have a total key in the grid. If not error.
-        if spectra_name not in list(grid.spectra.keys()):
-            raise exceptions.MissingSpectraType(
-                "The Grid does not contain the key '%s'" % spectra_name
-            )
-
-        # get particle age masks
-        mask = self._get_masks(young, old)
-
-        if np.sum(mask) == 0:
-            if verbose:
-                print("Age mask has filtered out all particles")
-
-            return np.zeros(len(grid.lam))
-
-        from ..extensions.particle_spectra import compute_particle_seds
-
-        # Prepare the arguments for the C function.
-        args = self._prepare_sed_args(
-            grid, fesc=fesc, spectra_type=spectra_name, mask=mask
-        )
-
-        # Get the integrated spectra in grid units (erg / s / Hz)
-        spec = compute_particle_seds(*args)
-
-        return spec
-
-    def _get_masks(self, young=None, old=None):
-        """
-        Get masks for which components we are handling, if a sub-component
-        has not been requested it's necessarily all particles.
-
-        Args:
-            young (float):
-                age in Myr at which to filter for young star particles
-            old (float):
-                age in Myr at which to filter for old star particles
-
-        Raises:
-            InconsistentParameter
-                Can't select for both young and old components
-                simultaneously
-
-        """
-
-        if young and old:
-            raise exceptions.InconsistentParameter(
-                "Galaxy sub-component can not be simultaneously young and old"
-            )
-        if young:
-            s = self.stars.log10ages <= np.log10(young)
-        elif old:
-            s = self.stars.log10ages > np.log10(old)
-        else:
-            s = np.ones(self.stars.nparticles, dtype=bool)
-
-        return s
-
     def get_line_los():
         """
         ParticleGalaxy specific method for obtaining the line luminosities
@@ -564,246 +356,6 @@ class Galaxy(BaseGalaxy):
         """
 
         pass
-
-    def get_particle_spectra_linecont(
-        self, grid, fesc=0.0, fesc_LyA=1.0, young=False, old=False
-    ):
-        """
-        Generate the line contribution spectra. This is only invoked if
-        fesc_LyA < 1.
-        """
-
-        # generate contribution of line emission alone and reduce the
-        # contribution of Lyman-alpha
-        linecont = self.generate_particle_lnu(
-            grid, spectra_name="linecont", old=old, young=young
-        )
-
-        # multiply by the Lyamn-continuum escape fraction
-        linecont *= 1 - fesc
-
-        # get index of Lyman-alpha
-        idx = grid.get_nearest_index(1216.0, grid.lam)
-        linecont[idx] *= fesc_LyA  # reduce the contribution of Lyman-alpha
-
-        return linecont
-
-    def get_particle_spectra_incident(
-        self, grid, young=False, old=False, label="", update=True
-    ):
-        """
-        Generate the incident (equivalent to pure stellar for stars) spectra
-        using the provided Grid.
-
-        Args:
-            grid (obj):
-                spectral grid object
-            update (bool):
-                flag for whether to update the `stellar` spectra
-                inside the galaxy object `spectra` dictionary.
-                These are the combined values of young and old.
-            young (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for young star particles
-            old (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for old star particles
-
-        Returns:
-            An Sed object containing the stellar spectra
-        """
-
-        lnu = self.generate_particle_lnu(grid, "incident", young=young, old=old)
-
-        sed = Sed(grid.lam, lnu)
-
-        if update:
-            self.spectra_array[label + "incident"] = sed
-
-        return sed
-
-    def get_particle_spectra_transmitted(
-        self, grid, fesc=0.0, young=False, old=False, label="", update=True
-    ):
-        """
-        Generate the transmitted spectra using the provided Grid. This is the
-        emission which is transmitted through the gas as calculated by the
-        photoionisation code.
-
-        Args:
-            grid (obj):
-                spectral grid object
-            fesc (float):
-                fraction of stellar emission that escapeds unattenuated from
-                the birth cloud (defaults to 0.0)
-            update (bool):
-                flag for whether to update the `stellar` spectra
-                inside the galaxy object `spectra` dictionary.
-                These are the combined values of young and old.
-            young (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for young star particles
-            old (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for old star particles
-
-        Returns:
-            An Sed object containing the transmitted spectra
-        """
-
-        lnu = (1.0 - fesc) * self.generate_particle_lnu(
-            grid, "transmitted", young=young, old=old
-        )
-
-        sed = Sed(grid.lam, lnu)
-
-        if update:
-            self.spectra_array[label + "transmitted"] = sed
-
-        return sed
-
-    def get_particle_spectra_nebular(
-        self, grid, fesc=0.0, update=True, young=False, label="", old=False
-    ):
-        """
-        Generate nebular spectra from a grid object and star particles.
-        The grid object must contain a nebular component.
-
-        Args:
-            grid (obj):
-                spectral grid object
-            fesc (float):
-                fraction of stellar emission that escapeds unattenuated from
-                the birth cloud (defaults to 0.0)
-            update (bool):
-                flag for whether to update the `intrinsic` and `attenuated`
-                spectra inside the galaxy object `spectra` dictionary.
-                These are the combined values of young and old.
-            young (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for young star particles
-            old (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for old star particles
-
-        Returns:
-            An Sed object containing the nebular spectra
-        """
-
-        lnu = self.generate_particle_lnu(grid, "nebular", young=young, old=old)
-
-        lnu *= 1 - fesc
-
-        sed = Sed(grid.lam, lnu)
-
-        if update:
-            self.spectra_array[label + "nebular"] = sed
-
-        return sed
-
-    def get_particle_spectra_reprocessed(
-        self,
-        grid,
-        fesc=0.0,
-        fesc_LyA=1.0,
-        young=False,
-        old=False,
-        label="",
-        update=True,
-    ):
-        """
-        Generates the intrinsic spectra, this is the sum of the escaping
-        radiation (if fesc>0), the transmitted emission, and the nebular
-        emission. The transmitted emission is the emission that is
-        transmitted through the gas. In addition to returning the intrinsic
-        spectra this saves the incident, nebular, and escaped spectra if
-        update is set to True.
-
-        Args:
-            grid (obj):
-                spectral grid object
-            fesc (float):
-                fraction of stellar emission that escapeds unattenuated from
-                the birth cloud (defaults to 0.0)
-            fesc_LyA (float):
-                fraction of Lyman-alpha emission that can escape unimpeded
-                by the ISM/IGM.
-            update (bool):
-                flag for whether to update the `intrinsic`, `stellar` and
-                `nebular` spectra inside the galaxy object `spectra`
-                dictionary. These are the combined values of young and old.
-            young (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for young star particles
-            old (bool, float):
-                if not False, specifies age in Myr at which to filter
-                for old star particles
-
-        Updates:
-            incident
-            transmitted
-            nebular
-            reprocessed
-            intrinsic
-
-            if fesc>0:
-                escaped
-
-        Returns:
-            An Sed object containing the intrinsic spectra.
-        """
-
-        # the incident emission
-        incident = self.get_particle_spectra_incident(
-            grid, update=update, young=young, old=old, label=label
-        )
-
-        # the emission which escapes the gas
-        if fesc > 0:
-            escaped = Sed(grid.lam, fesc * incident._lnu)
-
-        # the stellar emission which **is** reprocessed by the gas
-        transmitted = self.get_particle_spectra_transmitted(
-            grid, fesc, update=update, young=young, old=old, label=label
-        )
-        # the nebular emission
-        nebular = self.get_particle_spectra_nebular(
-            grid, fesc, update=update, young=young, old=old, label=label
-        )
-
-        # if the Lyman-alpha escape fraction is <1.0 suppress it.
-        if fesc_LyA < 1.0:
-            # get the new line contribution to the spectrum
-            linecont = self.get_particle_spectra_linecont(
-                grid, fesc=fesc, fesc_LyA=fesc_LyA
-            )
-
-            # get the nebular continuum emission
-            nebular_continuum = self.generate_particle_lnu(
-                grid, "nebular_continuum", young=young, old=old
-            )
-            nebular_continuum *= 1 - fesc
-
-            # redefine the nebular emission
-            nebular._lnu = linecont + nebular_continuum
-
-        # the reprocessed emission, the sum of transmitted, and nebular
-        reprocessed = nebular + transmitted
-
-        # the intrinsic emission, the sum of escaped, transmitted, and nebular
-        # if escaped exists other its simply the reprocessed
-        if fesc > 0:
-            intrinsic = reprocessed + escaped
-        else:
-            intrinsic = reprocessed
-
-        if update:
-            if fesc > 0:
-                self.spectra_array[label + "escaped"] = escaped
-            self.spectra_array[label + "reprocessed"] = reprocessed
-            self.spectra_array[label + "intrinsic"] = intrinsic
-
-        return reprocessed
 
     def get_particle_line_intrinsic(self, grid):
         # """
